+++ backend/app/data_output_configuration/__init__.py
@@ -2,6 +2,7 @@ from .coder.schema import CoderPlugin
 from .conveyor.schema import ConveyorPlugin
 from .databricks.schema import DatabricksTechnicalAssetConfiguration
 from .glue.schema import GlueTechnicalAssetConfiguration
+from .quarto.schema import QuartoPlugin
 from .redshift.schema import RedshiftTechnicalAssetConfiguration
 from .s3.schema import S3TechnicalAssetConfiguration
 from .snowflake.schema import SnowflakeTechnicalAssetConfiguration
@@ -11,6 +12,7 @@ __all__ = [
     "ConveyorPlugin",
     "DatabricksTechnicalAssetConfiguration",
     "GlueTechnicalAssetConfiguration",
+    "QuartoPlugin",
     "RedshiftTechnicalAssetConfiguration",
     "SnowflakeTechnicalAssetConfiguration",
     "S3TechnicalAssetConfiguration",