+++ backend/app/data_output_configuration/coder/schema.py
@@ -1,9 +1,14 @@
-from typing import ClassVar
+from typing import ClassVar, Optional
+from uuid import UUID
+
+from sqlalchemy.orm import Session
 
 from app.data_output_configuration.base_schema import (
     AssetProviderPlugin,
     PlatformMetadata,
 )
+from app.data_products.model import DataProduct as DataProductModel
+from app.users.schema import User
 
 
 class CoderPlugin(AssetProviderPlugin):
@@ -18,3 +23,15 @@ class CoderPlugin(AssetProviderPlugin):
         has_environments=False,
         detailed_name="VS Code - Coder",
     )
+
+    @classmethod
+    def get_url(
+        cls, id: UUID, db: Session, actor: User, environment: Optional[str] = None
+    ) -> str:
+        data_product = db.get(DataProductModel, id)
+        return f"http://localhost:8443/?folder=/home/coder/workspace/products/{data_product.namespace}"
+
+    @classmethod
+    def only_tile(cls) -> bool:
+        """Conveyor is currently only shown as a tile in the UI, as it doesn't have technical assets or detailed configuration options."""
+        return True