+++ backend/app/data_output_configuration/quarto/schema.py
@@ -3,7 +3,6 @@ from uuid import UUID
 
 from sqlalchemy.orm import Session
 
-from app.core.conveyor.notebook_builder import CONVEYOR_SERVICE
 from app.data_output_configuration.base_schema import (
     AssetProviderPlugin,
     PlatformMetadata,
@@ -12,18 +11,17 @@ from app.data_products.model import DataProduct as DataProductModel
 from app.users.schema import User
 
 
-class ConveyorPlugin(AssetProviderPlugin):
-    name: ClassVar[str] = "ConveyorPlatform"
+class QuartoPlugin(AssetProviderPlugin):
+    name: ClassVar[str] = "QuartoPlatform"
     version: ClassVar[str] = "1.0"
 
     _platform_metadata = PlatformMetadata(
-        display_name="Conveyor",
-        icon_name="conveyor-logo.svg",
-        platform_key="conveyor",
+        display_name="Quarto",
+        icon_name="quarto-logo.svg",
+        platform_key="quarto",
         parent_platform=None,
         has_environments=False,
-        detailed_name="Conveyor",
-        show_in_form=False,
+        detailed_name="Quarto - Documentation",
     )
 
     @classmethod
@@ -31,9 +29,4 @@ class ConveyorPlugin(AssetProviderPlugin):
         cls, id: UUID, db: Session, actor: User, environment: Optional[str] = None
     ) -> str:
         data_product = db.get(DataProductModel, id)
-        return CONVEYOR_SERVICE.generate_ide_url(data_product.namespace)
-
-    @classmethod
-    def only_tile(cls) -> bool:
-        """Conveyor is currently only shown as a tile in the UI, as it doesn't have technical assets or detailed configuration options."""
-        return True
+        return f"http://localhost:8443/?folder=/home/coder/workspace/products/{data_product.namespace}"