+++ frontend/src/pages/product-studio/product-studio-tabkeys.ts
@@ -1,0 +1,4 @@
+export enum TabKeys {
+    DataProducts = 'data-products',
+    OutputPorts = 'output-ports',
+}